Normalize declarations so that size_t is always unsigned long

+++ proofs/cbmc/include/ILP32/stddef.h
@@ -7,9 +7,9 @@
 #define _STDDEF_H 1
 
 /* "ILP32" stands for "Int, Long, and Pointer are 32-bit", so
- *  size_t is "unsigned int" which has the same size as a "pointer"
+ *  size_t is "unsigned long" which has the same size as a "pointer"
  */
-typedef unsigned int size_t;
+typedef unsigned long size_t;
 
 #define NULL (0)
